+++ CHANGELOG.md
@@ -5,11 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to 
 [Semantic Versioning](http://semver.org/).
 
-## 0.2.0.dev (development stage/unreleased/unstable)
+## 0.3.0.dev (development stage/unreleased/unstable)
+
+## 0.3.0
+### Added
+- Parameter `test` to lib and cli. Supported mode is "notification" to test email and telegram notifications.
+- Parameter `print_notificatons`. If True the lib is printing user friendly information to terminal. 
+### Fixed
+- Control notification settings to avoid exceptions
+
+## 0.2.0
 CLI update
 
 ## 0.1.2
-General updaes
+General updates
 
 ## 0.1.1
 General updates